split components, fix routing [WIP]
