+++ src/pages/Favourites.js
@@ -2,9 +2,12 @@ import React, { useState } from "react";
 import styled from "styled-components";
 import Header from "../components/Header";
 import Title from "../components/Title";
-import BadgeButtonClose from "../components/BadgeButtonClose";
-import { Link } from "react-router-dom";
-import { getFavouritesFromStorage } from "../api/storage";
+import FavouriteList from "../components/FavouriteList";
+// import { Link } from "react-router-dom";
+import {
+  getFavouritesFromStorage,
+  removeFavouritesFromStorage
+} from "../api/storage";
 
 const WrapperDiv = styled.div`
   display: flex;
@@ -12,23 +15,37 @@ const WrapperDiv = styled.div`
   align-items: center;
 `;
 
-const BadgeButtonList = styled.div`
-  max-width: 500px;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-`;
+// const BadgeButtonList = styled.div`
+//   max-width: 500px;
+//   display: flex;
+//   flex-direction: row;
+//   flex-wrap: wrap;
+// `;
 
-const StyledLink = styled(Link)`
-  text-decoration: none;
-  color: #fff;
-`;
+// const StyledLink = styled(Link)`
+//   text-decoration: none;
+//   color: #fff;
+// `;
 
-export default function Favourites({ id }) {
+export default function Favourites() {
   const [favourites, setFavourites] = useState(getFavouritesFromStorage());
 
-  function handleRemoveFavourite(favourites) {
-    setFavourites(favourites.filter(favourite => favourite.id === id));
+  // function handleChangeFavourites(favourite) {
+  //   const existingIndex = favourites.findIndex(
+  //     existingFavourite => existingFavourite.id === favourite.id
+  //   );
+  //   if (existingIndex === -1) {
+  //     const favouritesList = [...favourites, favourite];
+  //     setFavourites(favouritesList);
+  //   } else {
+  //     const favouritesList = [...favourites];
+  //     favouritesList.splice(existingIndex, 1);
+  //     setFavourites(favouritesList);
+  //   }
+  // }
+
+  function handleRemoveFavourite() {
+    removeFavouritesFromStorage();
   }
   console.log(favourites);
 
@@ -36,15 +53,12 @@ export default function Favourites({ id }) {
     <WrapperDiv>
       <Header showBubble children="Find your favourites here!" />
       <Title headline>FAVOURITES</Title>
-      <BadgeButtonList>
-        {favourites.map(favourite => (
-          <BadgeButtonClose key={favourite.id} onClick={handleRemoveFavourite}>
-            <StyledLink to={`${favourite.url}#${favourite.name.toLowerCase()}`}>
-              {favourite.name.toUpperCase()}
-            </StyledLink>
-          </BadgeButtonClose>
-        ))}
-      </BadgeButtonList>
+      <FavouriteList onClick={handleRemoveFavourite} />
     </WrapperDiv>
   );
 }
+
+// function handleRemoveFavourite() {
+//   setFavourites(favourites.filter(favourite => favourite.id === id));
+// }
+// console.log(favourites);
